requirements.txt
torch
torchvision
numpy
PyYAML
tqdm
Pillow
opencv-python
